+++ Makefile
@@ -1,5 +1,5 @@
+
 build:
 	go build -o gateway
 run:
-
 	./gateway
